Desk Set Structure
struct_deskSet is functioning a lot closer to what I had originally imagined. The user is now able to create a grid of desks simply by creating the structure object. The user can add or remove desks with the .addDesk() or .removeDesk functions and the structure will automatically create and place the desk objects accordingly. Only the .addDesk() function has been properly implemented and tested.